requirements.txt
langchain
langchain-community
langchain-groq
langchain-openai
langchain-chroma
python-dotenv
streamlit
langchain-text-splitters
langgraph
IPython